// src/sunat/nubefact.types.ts
export interface NubefactComprobanteResponse {
  tipo_de_comprobante: number;
  serie: string;
  numero: number;
  enlace: string;
  enlace_del_pdf: string;
  enlace_del_xml: string;
  enlace_del_cdr: string;
  aceptada_por_sunat: boolean;
  sunat_description: string | null;
  sunat_note: string | null;
  sunat_responsecode: string | null;
  sunat_soap_error: string;
  cadena_para_codigo_qr: string;
  codigo_hash: string;
}

export interface NubefactAnulacionResponse {
  numero: number;
  enlace: string;
  sunat_ticket_numero: string;
  aceptada_por_sunat: boolean;
  sunat_description: string | null;
  sunat_note: string | null;
  sunat_responsecode: string | null;
  sunat_soap_error: string;
  enlace_del_pdf: string;
  enlace_del_xml: string;
  enlace_del_cdr: string;
}